thread: Avoid the use of the potentially unsupported sem_init() function
sem_init() is deprecated on OS X, and it actually fails with ENOSYS. Using our
wrapped semaphore object is not an option, as it relies on the thread cleanup
that we can't rely on at this stage.

Instead, we could use a condvar to signal create completion. However, this
is more complicated than required; we can just hold the thread mutex until
thread creation is ready, and test for that lock in the thread main function.