vici: Honor DESTDIR when installing Python Egg
If `--with-pythoneggdir` is not set we fall back to the site-packages
directory as determined by the AM_PATH_PYTHON macro.

We also set PYTHONPATH explicitly so easy_install is happy if DESTDIR is set.

WIP: The problem with this is that easy_install creates or modifies a file
called easy-install.pth in the installation directory. So it's not actually
possible to simply copy the results in DESTDIR over to the actual system as
that file would have to be merged with any existing one.

 #914