utils: Add malloc/free wrappers returning aligned data
While we could use posix_memalign(3), that is not fully portable. Further, it
might be difficult on some platforms to properly catch it, which results in
invalid free()s when releasing such memory.

We instead use a simple wrapper, which allocates larger data, and saves the
pointer in the allocated header. This requires that memory is released using
a dedicated function.